pyproject.toml
[project]
name = "danika"
version = "0.1.0"
description = "A Discord bot to roll dice, provide information for various TTRPG systems, and fun."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiohttp>=3.11",
    "beautifulsoup4>=4.13",
    "discord-py>=2.7",
    "lark>=1.2",
    "loguru>=0.7",
    "pydantic>=2.12",
    "pydantic-settings>=2.13",
]

[dependency-groups]
dev = [
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
]

[tool.ruff]
target-version = "py314"
line-length = 100

[tool.ruff.lint]
select = [
    "F",     # pyflakes
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "I",     # isort
    "N",     # pep8-naming
    "UP",    # pyupgrade
    "B",     # flake8-bugbear
    "SIM",   # flake8-simplify
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short"
